requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
pystray==0.19.4
pillow==10.1.0
plyer==2.1.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
chromadb==0.4.18
sentence-transformers==2.2.2
openai==1.3.5
anthropic==0.7.7
mcp==0.6.0
cryptography==41.0.7
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.25.2
pyinstaller==6.2.0